Move response handling into callApi.
