+++ packages/core/shared/call-api.ts
@@ -1,4 +1,7 @@
-import { Message } from './types';
+import { nanoid } from 'nanoid';
+import { parseComplexResponse } from '../react/parse-complex-response';
+import { FunctionCall, JSONValue, Message } from './types';
+import { COMPLEX_HEADER, createChunkDecoder } from './utils';
 
 export async function callApi({
   api,
@@ -6,18 +9,24 @@ export async function callApi({
   body,
   credentials,
   headers,
-  signal,
+  abortController,
+  appendMessage,
   restoreMessagesOnFailure,
   onResponse,
+  onUpdate,
+  onFinish,
 }: {
   api: string;
   messages: Omit<Message, 'id'>[];
   body: Record<string, any>;
   credentials?: RequestCredentials;
   headers?: HeadersInit;
-  signal?: AbortSignal | null;
+  abortController?: () => AbortController | null;
   restoreMessagesOnFailure: () => void;
+  appendMessage: (message: Message) => void;
   onResponse?: (response: Response) => void | Promise<void>;
+  onUpdate: (merged: Message[], data: JSONValue[] | undefined) => void;
+  onFinish?: (message: Message) => void;
 }) {
   const response = await fetch(api, {
     method: 'POST',
@@ -26,7 +35,7 @@ export async function callApi({
       ...body,
     }),
     headers,
-    signal,
+    signal: abortController?.()?.signal,
     credentials,
   }).catch(err => {
     restoreMessagesOnFailure();
@@ -52,8 +61,84 @@ export async function callApi({
     throw new Error('The response body is empty.');
   }
 
-  return {
-    response,
-    reader: response.body.getReader(),
-  };
+  const reader = response.body.getReader();
+  const isComplexMode = response.headers.get(COMPLEX_HEADER) === 'true';
+
+  if (isComplexMode) {
+    const prefixMap = await parseComplexResponse({
+      reader,
+      abortControllerRef:
+        abortController != null ? { current: abortController() } : undefined,
+      update: onUpdate,
+    });
+
+    const responseMessages: Message[] = [];
+    const responseData: any = [];
+    for (const [type, item] of Object.entries(prefixMap)) {
+      if (onFinish && type === 'text') {
+        onFinish(item as Message);
+      }
+      if (type === 'data') {
+        responseData.push(item);
+      } else {
+        responseMessages.push(item as Message);
+      }
+    }
+
+    return { messages: responseMessages, data: responseData };
+  } else {
+    const createdAt = new Date();
+    const decode = createChunkDecoder(false);
+
+    // TODO-STREAMDATA: Remove this once Strem Data is not experimental
+    let streamedResponse = '';
+    const replyId = nanoid();
+    let responseMessage: Message = {
+      id: replyId,
+      createdAt,
+      content: '',
+      role: 'assistant',
+    };
+
+    // TODO-STREAMDATA: Remove this once Strem Data is not experimental
+    while (true) {
+      const { done, value } = await reader.read();
+      if (done) {
+        break;
+      }
+      // Update the chat state with the new message tokens.
+      streamedResponse += decode(value);
+
+      if (streamedResponse.startsWith('{"function_call":')) {
+        // While the function call is streaming, it will be a string.
+        responseMessage['function_call'] = streamedResponse;
+      } else {
+        responseMessage['content'] = streamedResponse;
+      }
+
+      appendMessage({ ...responseMessage });
+
+      // The request has been aborted, stop reading the stream.
+      if (abortController?.() === null) {
+        reader.cancel();
+        break;
+      }
+    }
+
+    if (streamedResponse.startsWith('{"function_call":')) {
+      // Once the stream is complete, the function call is parsed into an object.
+      const parsedFunctionCall: FunctionCall =
+        JSON.parse(streamedResponse).function_call;
+
+      responseMessage['function_call'] = parsedFunctionCall;
+
+      appendMessage({ ...responseMessage });
+    }
+
+    if (onFinish) {
+      onFinish(responseMessage);
+    }
+
+    return responseMessage;
+  }
 }
